+++ dipy/tracking/local/__init__.py
@@ -2,7 +2,15 @@ from .localtracking import LocalTracking
 from .tissue_classifier import (ActTissueClassifier, BinaryTissueClassifier,
                                 ThresholdTissueClassifier, TissueClassifier)
 from .direction_getter import DirectionGetter
+from dipy.align import Bunch
 from dipy.tracking import utils
 
 __all__ = ["LocalTracking", "ActTissueClassifier",
            "BinaryTissueClassifier", "ThresholdTissueClassifier"]
+
+# enum TissueClass (tissue_classifier.pxd) is not accessible
+# from here. To be changed when minimal cython version > 0.21.
+# cython 0.21 - cpdef enum to export values into Python-level namespace
+# https://github.com/cython/cython/commit/50133b5a91eea348eddaaad22a606a7fa1c7c457
+
+TissueTypes = Bunch(OUTSIDEIMAGE=-1, INVALIDPOINT=0, TRACKPOINT=1, ENDPOINT=2)
